LayoutTests/fast/frames/seamless/seamless-sandbox-flag-expected.txt

Test that sandbox disables nested seamless iframes correctly.
FAIL iframe.seamless should be true (of type boolean). Was undefined (of type undefined).
FAIL nestedFrame.seamless should be true (of type boolean). Was undefined (of type undefined).
FAIL window.getComputedStyle(iframe).width should be 200px. Was 300px.
PASS window.getComputedStyle(iframe).height is "150px"
